+++ playwright.config.ts
@@ -13,7 +13,7 @@ import { defineConfig, devices } from '@playwright/test';
  */
 export default defineConfig({
   testDir: './packages/ui/components',
-  testMatch: '*.e2e.test.js',
+  testMatch: '*.e2e.spec.js',
   /* Run tests in files in parallel */
   fullyParallel: true,
   /* Fail the build on CI if you accidentally left test.only in the source code. */
